+++ src/Engine/Blob.h
@@ -5,17 +5,17 @@ namespace Destiny
 	class Blob
 	{ 
 	public:
-		Blob(size_t length);
+		Blob(unsigned long long length);
 		~Blob();
 	public:
 		void*	getData();
-		size_t	getLength();
+		unsigned long long	getLength();
 	public:
-		void	copyfrom(void* data, size_t length);
+		void	copyfrom(void* data, unsigned long long length);
 		void	memset0();
 	private:
 		void*	m_data{ nullptr };
-		size_t	m_length{ 0 };
+		unsigned long long	m_length{ 0 };
 	};
 
 	inline void* Blob::getData()
@@ -23,7 +23,7 @@ namespace Destiny
 		return m_data;
 	}
 
-	inline size_t Blob::getLength()
+	inline unsigned long long Blob::getLength()
 	{
 		return m_length;
 	}
